+++ CHANGELOG.rst
@@ -6,7 +6,10 @@ Brewtils Changelog
 TBD
 
 - Plugin will periodically monitor if required dependencies are running to update status accordingly
+- Apply MD5 Check Sum of chunked files to ensure files are loaded into memory properly
+- Updated Plugin `max_concurrent` to support -1 to utilize the default formula that `concurrent.futures.ThreadPoolExecutor` supports `min(32, os.cpu_count() + 4)`
 - Updated SystemClient to utilize the local Garden name for default Namespace if none can be determined
+- Updated default Garden version to `UNKNOWN`
 
 3.27.2
 ------